requirements.txt
flask==2.3.2
line-bot-sdk==3.1.0
python-dotenv==1.0.0
ollama==0.1.4
torch==2.1.0
transformers==4.35.0
pillow==10.1.0
opencv-python==4.8.1.78
numpy==1.24.3
requests==2.31.0
